feat(compiler): export database storage in a more accessible way
allows to build on apollo-compiler's storage in separate databases. For example, it's now possible to

```rust
use apollo_compiler::{database::{AstStorage, DocumentStorage}};

#[salsa::database(AstStorage, DoumentStorage)]
pub struct AnotherDatabase {
    pub storage: salsa::Storage<AnotherDatabase>,
}
```
